Separate MosaicGridCompat into multiple files
